Move installation to utils
